# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Malmö Resturant Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/itemimages/unreport/{id}": {
            "put": {
                "tags": [
                    "Item Images"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ItemImagesController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/itemreviews/unreport/{id}": {
            "put": {
                "tags": [
                    "Item Reviews"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ItemReviewsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/items/new": {
            "post": {
                "tags": [
                    "Items"
                ],
                "description": "Create a new Item",
                "operationId": "App\\Http\\Controllers\\ItemsController::create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of Item",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of Item",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Price of Item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "resturant_id",
                        "in": "query",
                        "description": "Id of Resturant it belongs to.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Create a new resturant"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "description": "Get a specific item",
                "operationId": "App\\Http\\Controllers\\ItemsController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get a specific item"
                    }
                }
            }
        },
        "/api/items/delete/{id}": {
            "delete": {
                "tags": [
                    "Items"
                ],
                "description": "Delete a specific item",
                "operationId": "App\\Http\\Controllers\\ItemsController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Delete a specific item"
                    }
                }
            }
        },
        "/api/items/report/{id}": {
            "put": {
                "tags": [
                    "Items"
                ],
                "description": "Report an item",
                "operationId": "App\\Http\\Controllers\\ItemsController::report",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Report an item"
                    }
                }
            }
        },
        "/api/items/unreport/{id}": {
            "put": {
                "tags": [
                    "Items"
                ],
                "description": "unreport an item",
                "operationId": "App\\Http\\Controllers\\ItemsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item"
                    }
                }
            }
        },
...
        "/api/resturantimages/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturant Images"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ResturantImagesController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/resturantreviews/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturant Reviews"
                ],
                "description": "unreport an Resturant Review",
                "operationId": "App\\Http\\Controllers\\ResturantReviewsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an Resturant Review"
                    }
                }
            }
        },
...
        "/api/resturants/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturants"
                ],
                "description": "unreport a Resturant",
                "operationId": "App\\Http\\Controllers\\ResturantsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of resturant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport a Resturant"
                    }
                }
            }
        }
    }
}